requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
Django==1.8.3
django-celery==3.2.2
django-precise-bbcode==1.0.9
kombu==3.0.37
Pillow==2.9.0
pytz==2018.4
South==1.0.2
vine==1.1.4
